requirements.txt
asgiref==3.5.2
Django==3.2.15
django-extensions==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
